refactor: move Match UI state and ViewModel to match package
