--- app/src/main/java/com/karrar/movieapp/ui/match/MatchScreenViewModel.kt
@@ -1,8 +1,0 @@
-package com.karrar.movieapp.ui.match
-
-import androidx.lifecycle.ViewModel
-import dagger.hilt.android.lifecycle.HiltViewModel
-import javax.inject.Inject
-
-@HiltViewModel
-class MatchScreenViewModel @Inject constructor() : ViewModel()
