+++ app/src/main/java/com/karrar/movieapp/ui/match/result/MoviePagerAdapter.kt
@@ -1,10 +1,8 @@
 package com.karrar.movieapp.ui.match.result
 
 import com.karrar.movieapp.R
-import com.karrar.movieapp.domain.enums.HomeItemsType
 import com.karrar.movieapp.ui.base.BaseAdapter
 import com.karrar.movieapp.ui.base.BaseInteractionListener
-import com.karrar.movieapp.ui.match.questions.MatchQuestionUiState
 
 class MoviePagerAdapter(
     items: List<String>,
