+++ app/src/main/java/com/karrar/movieapp/ui/match/result/mappers/ActorUIStateMapper.kt
@@ -2,12 +2,12 @@ package com.karrar.movieapp.ui.match.result.mappers
 
 import com.karrar.movieapp.domain.mappers.Mapper
 import com.karrar.movieapp.domain.models.Actor
-import com.karrar.movieapp.ui.match.questions.MatchQuestionUiState
+import com.karrar.movieapp.ui.match.MatchUiState
 import javax.inject.Inject
 
-class ActorUIStateMapper @Inject constructor() : Mapper<Actor, MatchQuestionUiState.ActorUiState> {
-    override fun map(input: Actor): MatchQuestionUiState.ActorUiState {
-        return MatchQuestionUiState.ActorUiState(
+class ActorUIStateMapper @Inject constructor() : Mapper<Actor, MatchUiState.ActorUiState> {
+    override fun map(input: Actor): MatchUiState.ActorUiState {
+        return MatchUiState.ActorUiState(
             id = input.actorID,
             imageUrl = input.actorImage,
             name = input.actorName,
