+++ app/src/main/java/com/karrar/movieapp/ui/match/result/mappers/CrewUIStateMapper.kt
@@ -2,12 +2,12 @@ package com.karrar.movieapp.ui.match.result.mappers
 
 import com.karrar.movieapp.domain.mappers.Mapper
 import com.karrar.movieapp.domain.models.Crew
-import com.karrar.movieapp.ui.match.questions.MatchQuestionUiState
+import com.karrar.movieapp.ui.match.MatchUiState
 import javax.inject.Inject
 
-class CrewUIStateMapper @Inject constructor() : Mapper<Crew, MatchQuestionUiState.CrewUIState> {
-    override fun map(input: Crew): MatchQuestionUiState.CrewUIState {
-        return MatchQuestionUiState.CrewUIState(
+class CrewUIStateMapper @Inject constructor() : Mapper<Crew, MatchUiState.CrewUIState> {
+    override fun map(input: Crew): MatchUiState.CrewUIState {
+        return MatchUiState.CrewUIState(
             name = input.crewMemberName,
             job = input.crewMemberJob
         )
