+++ app/src/main/java/com/karrar/movieapp/ui/match/result/mappers/MediaUIStateMapper.kt
@@ -2,12 +2,12 @@ package com.karrar.movieapp.ui.match.result.mappers
 
 import com.karrar.movieapp.domain.mappers.Mapper
 import com.karrar.movieapp.domain.models.Media
-import com.karrar.movieapp.ui.match.questions.MatchQuestionUiState
+import com.karrar.movieapp.ui.match.MatchUiState
 import javax.inject.Inject
 
-class MediaUIStateMapper @Inject constructor() : Mapper<Media, MatchQuestionUiState.MediaUiState> {
-    override fun map(input: Media): MatchQuestionUiState.MediaUiState {
-        return MatchQuestionUiState.MediaUiState(
+class MediaUIStateMapper @Inject constructor() : Mapper<Media, MatchUiState.MediaUiState> {
+    override fun map(input: Media): MatchUiState.MediaUiState {
+        return MatchUiState.MediaUiState(
             id = input.mediaID,
             imageUrl = input.mediaImage,
             mediaTitle = input.mediaName,
